backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
elasticsearch==8.13.0
python-chess==1.999
structlog==24.1.0
slowapi==0.1.9
httpx==0.27.0           # for async test client
pytest==8.2.0
pytest-asyncio==0.23.6
